+++ plugin/src/main/groovy/org/openbakery/XcodeBuildCleanTask.groovy
@@ -24,6 +24,7 @@ class XcodeBuildCleanTask extends DefaultTask {
 
 	XcodeBuildCleanTask() {
 		super()
+		dependsOn('keychain-clean', 'provisioning-clean', 'hockeykit-clean', 'testflight-clean', 'hockeyapp-clean')
 		this.description = "Cleans up the generated files from the previous build"
 	}