+++ plugin/src/main/groovy/org/openbakery/XcodePlugin.groovy
@@ -160,6 +160,7 @@ class XcodePlugin implements Plugin<Project> {
 
 			defineDynamicDependencies()
 		}
+
 	}
 
 	def void defineExtensions() {
@@ -174,6 +175,8 @@ class XcodePlugin implements Plugin<Project> {
 	def void defineTasks() {
 		project.task('keychain-create', type: KeychainCreateTask, group: XCODE_GROUP_NAME)
 		project.task('xcodebuild', type: XcodeBuildTask, group: XCODE_GROUP_NAME)
+		project.task('build', type: XcodeBuildTask, group: XCODE_GROUP_NAME)
+		project.task('test', type: XcodeTestTask, group: XCODE_GROUP_NAME)
 		project.task('infoplist-modify', type: InfoPlistModifyTask, group: XCODE_GROUP_NAME)
 		project.task('provisioning-install', type: ProvisioningInstallTask, group: XCODE_GROUP_NAME)
 		project.task('archive', type: XcodeBuildArchiveTask, group: XCODE_GROUP_NAME)
@@ -210,7 +213,7 @@ class XcodePlugin implements Plugin<Project> {
 		Task archive = project.tasks."archive"
 		archive.dependsOn("clean")
 
-		Task uiautomation = project.tasks.'uiautomation'
+		//Task uiautomation = project.tasks.'uiautomation'
 
 		Task hockeyKitManifest = project.tasks.'hockeykit-manifest'
 		Task hockeyKitArchiveTask = project.tasks.'hockeykit-archive'
@@ -219,28 +222,6 @@ class XcodePlugin implements Plugin<Project> {
 		Task hockey = project.tasks.'hockeykit'
 		hockey.dependsOn(hockeyKitArchiveTask, hockeyKitManifest, hockeyKitImageTask, hockeyKitNotes)
 		hockeyKitArchiveTask.dependsOn(archive)
-
-		Task keychainCleanup = project.tasks.'keychain-clean'
-		Task xcodebuildCleanup = project.tasks.'clean'
-		Task provisioningCleanup = project.tasks.'provisioning-clean'
-		Task hockeyKitCleanTask = project.tasks.'hockeykit-clean'
-		Task testFlightClean = project.tasks.'testflight-clean'
-		Task hockeyAppClean = project.tasks.'hockeyapp-clean'
-
-		xcodebuildCleanup.dependsOn(keychainCleanup)
-		xcodebuildCleanup.dependsOn(provisioningCleanup)
-		xcodebuildCleanup.dependsOn(hockeyKitCleanTask)
-		xcodebuildCleanup.dependsOn(testFlightClean);
-		xcodebuildCleanup.dependsOn(hockeyAppClean);
-		xcodebuildCleanup.dependsOn(hockeyAppClean);
-
-
-		xcodebuild.dependsOn(project.tasks.'keychain-create')
-		xcodebuild.dependsOn(project.tasks.'provisioning-install')
-		xcodebuild.dependsOn(infoplistModify)
-
-		Task codesign = project.tasks.'codesign'
-		archive.dependsOn(codesign)
 
 
 	}