+++ plugin/src/test/groovy/org/openbakery/XcodeBuildTaskTest.groovy
@@ -31,7 +31,11 @@ class XcodeBuildTaskTest {
 	List<String> expectedCommandList
 
 	String currentDir = new File('').getAbsolutePath()
+<<<<<<< HEAD
 	String projectDir
+=======
+    String projectDir
+>>>>>>> xcode5
 
 	@BeforeMethod
 	def setup() {
@@ -39,7 +43,11 @@ class XcodeBuildTaskTest {
 
 		project = ProjectBuilder.builder().build()
 		project.buildDir = new File('build').absoluteFile
+<<<<<<< HEAD
 		projectDir = project.projectDir.absolutePath
+=======
+        projectDir = project.projectDir.absolutePath
+>>>>>>> xcode5
 		project.apply plugin: org.openbakery.XcodePlugin
 
 		xcodeBuildTask = project.getTasks().getByPath('xcodebuild')
@@ -83,10 +91,10 @@ class XcodeBuildTaskTest {
 	}
 
 	def void addExpectedDefaultDirs() {
-		expectedCommandList.add("DSTROOT=" + currentDir + "/build/dst")
-		expectedCommandList.add("OBJROOT=" + currentDir + "/build/obj")
-		expectedCommandList.add("SYMROOT=" + currentDir + "/build/sym")
-		expectedCommandList.add("SHARED_PRECOMPS_DIR=" + currentDir + "/build/shared")
+		expectedCommandList.add("DSTROOT=" + currentDir + "${File.separator}build${File.separator}dst")
+		expectedCommandList.add("OBJROOT=" + currentDir + "${File.separator}build${File.separator}obj")
+		expectedCommandList.add("SYMROOT=" + currentDir + "${File.separator}build${File.separator}sym")
+		expectedCommandList.add("SHARED_PRECOMPS_DIR=" + currentDir + "${File.separator}build${File.separator}shared")
 	}
 
 	@Test
@@ -101,10 +109,10 @@ class XcodeBuildTaskTest {
 		expectedCommandList.add("Debug")
 
 
-		project.xcodebuild.dstRoot = new File(currentDir + '/mydst')
-		project.xcodebuild.objRoot = new File(currentDir + '/myobj')
-		project.xcodebuild.symRoot = new File(currentDir + '/mysym')
-		project.xcodebuild.sharedPrecompsDir = new File(currentDir + '/myshared')
+		project.xcodebuild.dstRoot = new File(currentDir + "${File.separator}mydst")
+		project.xcodebuild.objRoot = new File(currentDir + "${File.separator}myobj")
+		project.xcodebuild.symRoot = new File(currentDir + "${File.separator}mysym")
+		project.xcodebuild.sharedPrecompsDir = new File(currentDir + "${File.separator}myshared")
 
 		expectedCommandList.add("DSTROOT=" + project.xcodebuild.dstRoot.absolutePath)
 		expectedCommandList.add("OBJROOT=" + project.xcodebuild.objRoot.absolutePath)